Move existing catalog check.
