+++ src/hats_import/catalog/run_import.py
@@ -6,7 +6,6 @@ The actual logic of the map reduce is in the `map_reduce.py` file.
 
 import os
 import pickle
-from pathlib import Path
 
 import hats.io.file_io as io
 from hats.catalog import PartitionInfo
@@ -19,26 +18,12 @@ from hats_import.catalog.arguments import ImportArguments
 from hats_import.catalog.resume_plan import ResumePlan
 
 
-def _validate_arguments(args):
-    """
-    Verify that the args for run are valid: they exist, are of the appropriate type,
-    and do not specify an output which is a valid catalog.
-
-    Raises ValueError if they are invalid.
-    """
+def run(args, client):
+    """Run catalog creation pipeline."""
     if not args:
         raise ValueError("args is required and should be type ImportArguments")
     if not isinstance(args, ImportArguments):
         raise ValueError("args must be type ImportArguments")
-
-    potential_path = Path(args.output_path) / args.output_artifact_name
-    if is_valid_catalog(potential_path):
-        raise ValueError(f"Output path {potential_path} already contains a valid catalog")
-
-
-def run(args, client):
-    """Run catalog creation pipeline."""
-    _validate_arguments(args)
 
     resume_plan = ResumePlan(import_args=args)
 
@@ -137,7 +122,7 @@ def run(args, client):
 
     # All done - write out the metadata
     if resume_plan.should_run_finishing:
-        with resume_plan.print_progress(total=4, stage_name="Finishing") as step_progress:
+        with resume_plan.print_progress(total=5, stage_name="Finishing") as step_progress:
             partition_info = PartitionInfo.from_healpix(resume_plan.get_destination_pixels())
             partition_info_file = paths.get_partition_info_pointer(args.catalog_path)
             partition_info.write_to_file(partition_info_file)
@@ -151,12 +136,14 @@ def run(args, client):
             else:
                 partition_info.write_to_metadata_files(args.catalog_path)
             step_progress.update(1)
+            io.write_fits_image(raw_histogram, paths.get_point_map_file_pointer(args.catalog_path))
+            step_progress.update(1)
             catalog_info = args.to_table_properties(
                 total_rows, partition_info.get_highest_order(), partition_info.calculate_fractional_coverage()
             )
             catalog_info.to_properties_file(args.catalog_path)
             step_progress.update(1)
-            io.write_fits_image(raw_histogram, paths.get_point_map_file_pointer(args.catalog_path))
-            step_progress.update(1)
             resume_plan.clean_resume_files()
+            step_progress.update(1)
+            assert is_valid_catalog(args.catalog_path)
             step_progress.update(1)
